fix types + edit func savePost after  PR
